Fix a pre-existing MPI type-size mismatch in Dataset structure serialization (found via AddressSanitizer)
Dataset::sendStructure()/recvStructure() declared a temporary size
variable `ts` as `int` (4 bytes) but packed/unpacked it with the
MPI_SIZE_T datatype (8 bytes), three times over in each direction
(comment length, numAtomsPerElement size, atom count). This is
undefined behavior -- a stack-buffer-overflow read/write past a
4-byte variable -- on every structure sent to or received by any
non-rank-0 MPI process.

git blame traces this to n2p2's original v2.0.0 commit (9769786,
2018-10-29), years before this project's GPU port -- not a bug this
project introduced. Its visible impact depends on incidental stack
layout, consistent with being silent in the CPU build and
consequential in the GPU build, and invisible to single-rank runs
(rank 0 never calls sendStructure/recvStructure for its own
structures).

Found while tracing water's real nnp-train GPU divergence: an
instrumented per-update trace of the Kalman-filter weight-update path
showed a corrupted Training::countUpdates value that decoded cleanly
as a stray IEEE double, pointing to an out-of-bounds write rather
than uninitialized memory. compute-sanitizer found nothing (it only
covers CUDA/device memory); rebuilding with AddressSanitizer caught
this exact overflow immediately.

Fixed (int -> size_t, matching MPI_SIZE_T) and confirmed via
AddressSanitizer: 0 errors now, versus a caught overflow before.
This is a real, validated fix, kept on its own merits -- but water's
training divergence persists unchanged with the fix applied, so this
was not the (or not the whole) cause. That investigation continues
separately.

+++ src/libnnptrain/Dataset.cpp
@@ -253,7 +253,7 @@ int Dataset::sendStructure(Structure const& structure, int dest) const
     unsigned char*   buf = 0;         // Send buffer.
     int              bs  = 0;         // Send buffer size.
     int              p   = 0;         // Send buffer position.
-    int              ts  = 0;         // Size for temporary stuff.
+    size_t           ts  = 0;         // Size for temporary stuff.
     Structure const& s   = structure; // Shortcut for structure.
 
     bs = calculateBufferSize(s);
@@ -465,7 +465,7 @@ int Dataset::recvStructure(Structure* const structure, int src)
     unsigned char*   buf = 0;         // Receive buffer.
     int              bs  = 0;         // Receive buffer size.
     int              p   = 0;         // Receive buffer position.
-    int              ts  = 0;         // Size for temporary stuff.
+    size_t           ts  = 0;         // Size for temporary stuff.
     Structure* const s   = structure; // Shortcut for structure.
     MPI_Status       ms;
